requirements.txt
pycryptodome==3.17
PyQt5==5.15.9
